requirements.txt
art~=6.4
python-dotenv~=1.0.1
openai~=1.55.3
requests~=2.32.3